Added toast on error parsing payload from WalletConnect request

+++ app/src/main/java/com/concordium/wallet/data/walletconnect/Payload.kt
@@ -4,7 +4,7 @@ data class Payload(
     val address: ContractAddress,
     val amount: String,
     var maxEnergy: Long,
-    var maxContractExecutionEnergy: Int,
+    var maxContractExecutionEnergy: Long,
     val message: String,
     val receiveName: String
 )
